conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.5.1
annotated-types=0.7.0=pypi_0
anyio=4.12.0=pypi_0
asgiref=3.8.1=py313haa95532_0
brotlicffi=1.2.0.0=py313h885b0b7_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.12.2=haa95532_0
cachetools=6.2.4=pypi_0
certifi=2026.01.04=py313haa95532_0
cffi=2.0.0=py313h02ab6af_1
charset-normalizer=3.4.4=py313haa95532_0
click=8.2.1=py313haa95532_1
colorama=0.4.6=py313haa95532_0
distro=1.9.0=pypi_0
django=5.2=py313haa95532_0
django-debug-toolbar=6.1.0=pypi_0
djangorestframework=3.16.1=pyhd8ed1ab_0
expat=2.7.1=h8ddb27b_0
google-ai-generativelanguage=0.6.15=pypi_0
google-api-core=2.28.1=pypi_0
google-api-python-client=2.187.0=pypi_0
google-auth=2.45.0=pypi_0
google-auth-httplib2=0.3.0=pypi_0
google-genai=1.56.0=pypi_0
google-generativeai=0.8.6=pypi_0
googleapis-common-protos=1.72.0=pypi_0
grpcio=1.76.0=pypi_0
grpcio-status=1.71.2=pypi_0
h11=0.16.0=pypi_0
httpcore=1.0.9=pypi_0
httplib2=0.31.0=pypi_0
httpx=0.28.1=pypi_0
idna=3.11=py313haa95532_0
jiter=0.12.0=pypi_0
libffi=3.4.4=hd77b12b_1
libmpdec=4.0.0=h827c3e9_0
libzlib=1.3.1=h02ab6af_0
openai=2.14.0=pypi_0
openssl=3.0.18=h543e019_0
pillow=12.0.0=pypi_0
pip=25.2=pyhc872135_0
proto-plus=1.27.0=pypi_0
protobuf=5.29.5=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.2=pypi_0
pycparser=2.23=py313haa95532_0
pydantic=2.12.5=pypi_0
pydantic-core=2.41.5=pypi_0
pyparsing=3.3.1=pypi_0
pysocks=1.7.1=py313haa95532_1
python=3.13.7=h260b955_100_cp313
python-dotenv=1.2.1=pypi_0
python_abi=3.13=1_cp313
requests=2.32.5=py313haa95532_1
rsa=4.9.1=pypi_0
setuptools=78.1.1=py313haa95532_0
sniffio=1.3.1=pypi_0
sqlite=3.50.2=hda9a48d_1
sqlparse=0.5.2=py313haa95532_0
tenacity=9.1.2=pypi_0
tk=8.6.15=hf199647_0
tqdm=4.67.1=pypi_0
typing-inspection=0.4.2=pypi_0
typing_extensions=4.15.0=py313haa95532_0
tzdata=2025b=h04d1e81_0
ucrt=10.0.22621.0=haa95532_0
uritemplate=4.2.0=pypi_0
urllib3=2.6.1=py313haa95532_0
vc=14.3=h2df5915_10
vc14_runtime=14.44.35208=h4927774_10
vs2015_runtime=14.44.35208=ha6b5a95_10
waitress=3.0.2=py313haa95532_0
websockets=15.0.1=pypi_0
wheel=0.45.1=py313haa95532_0
win_inet_pton=1.1.0=py313haa95532_1
xz=5.6.4=h4754444_1
zlib=1.3.1=h02ab6af_0
